autosnapshot. rename example packagename rework test case to use startPage and auto navigate
