+++ selenium-automation-example/src/main/java/com/github/licanhua/example/expedia/pages/CarSearchComponent.java
@@ -1,4 +1,23 @@
-package framework.expedia.pages;
+/*
+ *
+ *  * Copyright (C) 2016 The Selenium Automation Framework Authors
+ *  *
+ *  * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
+ *  * in compliance with the License. You may obtain a copy of the License at
+ *  *
+ *  * http://www.apache.org/licenses/LICENSE-2.0
+ *  *
+ *  * Unless required by applicable law or agreed to in writing, software distributed under the License
+ *  * is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
+ *  * or implied. See the License for the specific language governing permissions and limitations under
+ *  * the License.
+ *  *
+ *  * Created by canhua li (licanhua@live.com)
+ *  *
+ *
+ */
+
+package com.github.licanhua.example.expedia.pages;
 
 import com.github.licanhua.test.framework.Container;
 import com.github.licanhua.test.framework.annotation.OptionalElement;
@@ -9,7 +28,6 @@ import org.openqa.selenium.support.FindBy;
 
 import java.text.SimpleDateFormat;
 import java.util.Calendar;
-import java.util.Date;
 
 public class CarSearchComponent extends Container{
     private static Logger logger = Logger.getLogger(CarSearchComponent.class.getName());
