--- selenium-automation-example/src/main/java/framework/expedia/pages/HomePage.java
@@ -1,8 +1,0 @@
-package framework.expedia.pages;
-
-import com.github.licanhua.test.framework.Page;
-import org.openqa.selenium.support.FindBy;
-
-public class HomePage extends Page {
-
-}
