+++ selenium-automation-example/src/main/resources/config/com.github.licanhua.example.expedia.test.conf
@@ -1,0 +1,1 @@
+startPage : "https://www.expedia.com"
