+++ selenium-automation-example/src/main/resources/log4j.properties
@@ -1,5 +1,5 @@
 log4j.rootLogger=info, stdout
-log4j.logger.com.github.licanhua.test.framework=debug
+log4j.logger.com.github.licanhua.test.framework=info
 log4j.logger.org.apache.http.headers=off
 log4j.logger.org.apache.http.wire=off
