+++ selenium-automation-framework/src/main/resources/config/automation.properties
@@ -3,9 +3,13 @@ browserName=chrome
 # in com.github.licanhua.test.framework.base.WaitFunctions.java
 # waitDurationInSeconds should be big than ajaxTimeoutInSeconds
 # because it will call multiple findElement(s) until waitDurationInSeconds timeout.
-waitDurationInSeconds=30
+waitDurationInSeconds=60
 # ajaxTimeoutInSeconds: How many seconds a webElement.findElement(s) timeout. When you call any WebElement function
 # a findElements is trigged and Selenium will throw NoSuchElementException if ajaxTimeoutInSeconds reached.
 # ajaxTimeoutInSeconds should less than waitDurationInSeconds
-ajaxTimeoutInSeconds=10
+ajaxTimeoutInSeconds=20
 remoteWebDriverAddress=
+autoSnapshot=false
+targetEnvironment=Prod
+#the default page will automatically navigate to by framework
+startPage=
