+++ selenium-automation-framework/src/test/resources/log4j.properties
@@ -1,5 +1,5 @@
-log4j.rootLogger=debug, stdout
-log4j.logger.com.github.licanhua.test.framework=debug
+log4j.rootLogger=info, stdout
+log4j.logger.com.github.licanhua.test.framework=info
 log4j.logger.org.apache.http.headers=off
 log4j.logger.org.apache.http.wire=off
